src/services/gmail_test.py
# src/services/gmail_test.py
from google.oauth2.credentials import Credentials
from google_auth_oauthlib.flow import InstalledAppFlow
from google.auth.transport.requests import Request
from googleapiclient.discovery import build
import pickle
import os

SCOPES = ['https://www.googleapis.com/auth/gmail.readonly']


def test_gmail_connection():
    creds = None
    token_file = 'token.pickle'

    if os.path.exists(token_file):
        with open(token_file, 'rb') as token:
            creds = pickle.load(token)

    if not creds or not creds.valid:
        if creds and creds.expired and creds.refresh_token:
            creds.refresh(Request())
        else:
            flow = InstalledAppFlow.from_client_secrets_file(
                'config/credentials.json', SCOPES)
            creds = flow.run_local_server(port=0)

        with open(token_file, 'wb') as token:
            pickle.dump(creds, token)

    service = build('gmail', 'v1', credentials=creds)

    results = service.users().labels().list(userId='me').execute()
    labels = results.get('labels', [])

    print(f"Successfully connected! Found {len(labels)} labels:")
    for label in labels[:5]:
        print(f"  - {label['name']}")

    messages = service.users().messages().list(
        userId='me', maxResults=5).execute()

    print(f"\nFound {len(messages.get('messages', []))} recent messages")

    return True


if __name__ == "__main__":
    test_gmail_connection()
